Flush when activating plugin

+++ redis-cache.php
@@ -30,10 +30,11 @@ class RedisObjectCache {
 
 		load_plugin_textdomain( 'redis-cache', false, 'redis-cache/languages' );
 
-		register_activation_hook( __FILE__, array( $this, 'flush_cache_if_connected' ) );
-		register_deactivation_hook( __FILE__, array( $this, 'on_deactivation' ) );
+		register_activation_hook( __FILE__, 'wp_cache_flush' );
 
 		$this->page = is_multisite() ? 'settings.php?page=redis-cache' : 'options-general.php?page=redis-cache';
+
+		add_action( 'deactivate_plugin', array( $this, 'on_deactivation' ) );
 
 		add_action( is_multisite() ? 'network_admin_menu' : 'admin_menu', array( $this, 'add_admin_menu_page' ) );
 		add_action( 'admin_notices', array( $this, 'show_admin_notices' ) );
@@ -142,14 +143,6 @@ class RedisObjectCache {
 
 	}
 
-	public function flush_cache_if_connected() {
-
-		if ( $this->validate_object_cache_dropin() ) {
-			wp_cache_flush();
-		}
-
-	}
-
 	public function get_status() {
 
 		if ( ! $this->object_cache_dropin_exists() ) {
@@ -253,7 +246,6 @@ class RedisObjectCache {
 			switch ( $_GET[ 'message' ] ) {
 
 				case 'cache-enabled':
-					$this->flush_cache_if_connected();
 					$message = __( 'Object Cache enabled.', 'redis-cache' );
 					break;
 				case 'enable-cache-failed':
@@ -272,7 +264,6 @@ class RedisObjectCache {
 					$error = __( 'Object Cache could not be flushed.', 'redis-cache' );
 					break;
 				case 'dropin-updated':
-					$this->flush_cache_if_connected();
 					$message = __( 'Updated object cache drop-in and enabled Redis object cache.', 'redis-cache' );
 					break;
 				case 'update-dropin-failed':
@@ -321,7 +312,6 @@ class RedisObjectCache {
 							break;
 
 						case 'disable-cache':
-							$this->flush_cache_if_connected();
 							$result = $wp_filesystem->delete( WP_CONTENT_DIR . '/object-cache.php' );
 							$message = $result ? 'cache-disabled' : 'disable-cache-failed';
 							break;
@@ -379,12 +369,18 @@ class RedisObjectCache {
 
 	}
 
-	public function on_deactivation() {
+	public function on_deactivation( $plugin ) {
 
 		global $wp_filesystem;
 
-		if ( $this->validate_object_cache_dropin() && $this->initialize_filesystem( '', true ) ) {
-			$wp_filesystem->delete( WP_CONTENT_DIR . '/object-cache.php' );
+		if ( $plugin === plugin_basename( __FILE__ ) ) {
+
+			wp_cache_flush();
+
+			if ( $this->validate_object_cache_dropin() && $this->initialize_filesystem( '', true ) ) {
+				$wp_filesystem->delete( WP_CONTENT_DIR . '/object-cache.php' );
+			}
+
 		}
 
 	}
